% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{K-Means Clustering}
\usage{
k_means(x, centers, PCA = F, plus.plus = F)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix
(such as a numeric vector or a data frame with all numeric columns).}

\item{centers}{the number of clusters k.}

\item{PCA}{logical indicating if you want to preform a PCA transformation on the data.}

\item{plus.plus}{logical indicating if you want to use the k-means++ algorithm for
improved choosing of initial values.}
}
\value{
A list containing cluster means, clustering vector, and total SSE.
}
\description{
Preform k-means clustering on a dataset.
}
